+++ api_tests/entry_api_test.py
@@ -59,7 +59,7 @@ def test_read_no_bucket(base_url, session):
     """Should return 404 if no bucket found"""
     resp = session.get(f'{base_url}/b/xxx/entry?ts=100')
     assert resp.status_code == 404
-    assert 'xxx' in resp.headers["-x-reduct-error"]
+    assert 'xxx' in resp.headers["x-reduct-error"]
 
 
 def test_read_no_data(base_url, session, bucket):
@@ -74,14 +74,14 @@ def test_read_bad_ts(base_url, session, bucket):
     resp = session.get(f'{base_url}/b/{bucket}/entry?ts=XXXX')
 
     assert resp.status_code == 422
-    assert 'XXX' in resp.headers["-x-reduct-error"]
+    assert 'XXX' in resp.headers["x-reduct-error"]
 
 
 def test_read_bad_no_entry(base_url, session, bucket):
     """Should return 400 if ts is bad"""
     resp = session.get(f'{base_url}/b/{bucket}/entry')
     assert resp.status_code == 404
-    assert 'entry' in resp.headers["-x-reduct-error"]
+    assert 'entry' in resp.headers["x-reduct-error"]
 
 
 @requires_env("API_TOKEN")
@@ -105,14 +105,14 @@ def test_write_no_bucket(base_url, session):
     """Should return 404 if no bucket found"""
     resp = session.post(f'{base_url}/b/xxx/entry?ts=100')
     assert resp.status_code == 404
-    assert 'xxx' in resp.headers["-x-reduct-error"]
+    assert 'xxx' in resp.headers["x-reduct-error"]
 
 
 def test_write_bad_ts(base_url, session, bucket):
     """Should return 422 if ts is bad"""
     resp = session.post(f'{base_url}/b/{bucket}/entry?ts=XXXX')
     assert resp.status_code == 422
-    assert 'XXX' in resp.headers["-x-reduct-error"]
+    assert 'XXX' in resp.headers["x-reduct-error"]
 
 
 def test_get_record_ok(base_url, session, bucket):